==> 1. EDA_Datos_Originales.ipynb <==
# Análisis Exploratorio de Datos (EDA - Sobre los datos originales):

Se tiene a discposición 3 archivos en formato ".json" comprimidos (steam_games, user_reviews y user_items) guardados en la carpeta "Original_Data". En este análisis preliminar solo se abrirán los archivos, se revisará la estructura de los datos, que tipo de formato tiene cada columna y si hay registros vacíos. También se verificará que campos pueden usarse para relacionar a los diferentes datasets entre sí. 

### 1. Importación de librerías para abrir los documentos:

import gzip                             # Para descomprimir los archivos ".json"
import pandas as pd                     # Para convertir los archivos en dataframes
import json                             # Para leer los documentos ".json"
import ast                              # Para abrir los archivos con error en el formato ".json"
import chardet                          # Para encontrar el encoding de los archivos
import warnings                         # Para Ignorar errores
warnings.filterwarnings("ignore")                           

### 2. Fórmula para descomprimir y abrir los archivos:

Debido a que algunas líneas presentaban errores, no fue sencillo abrir leer los archivos así que se opta por una función que lea los archivos comprimidos ".json", ignores e imprima los errores.

# Función para abrir los archivos:

def abrir_comprimido(archivo):
    
    data_list = []
    with gzip.open(archivo, 'rt', encoding='utf-8') as file:
        for line in file.readlines():
            try:
                data = json.loads(line.strip())
                data_list.append(data)
            except json.JSONDecodeError as e:
                print(f"Error en línea: {line.strip()}")
                print(e)
    return data_list



### 3. Análisis Preliminar del archivo "steam_games.json.gz":

# Abrir archivo y convertirlo en datafame:

archivo = 'Original_Data\steam_games.json.gz'
df_Steam_Games = pd.DataFrame(abrir_comprimido(archivo)) # uso de la fórmula creada para descomprimir y leer archivo.

# Visualizar dataframe:

df_Steam_Games.head()

Se aprecia que las primeras líneas contienen registros vacíos. A continuación se procede a revisar cuantos registros pueden estar vacíos:

# Muestra la cuenta de NaN por columnas:
print(df_Steam_Games.isna().sum())

La idea inicial de este análisis preliminar no incluía imputación o eliminación de registros, pero debido a que representan la mayorías de los datos se procede a eliminar estos registros:

df_Steam_Games.dropna(how='all',inplace=True) # Eliminar registros que contienen vacíos en todas las columnas
print(df_Steam_Games.isna().sum()) # Muestra la cuenta de NaN por columnas

Se reduce ampliamente los registros del dataset, que eal parecer en su mayoría eran vacíos. A continuación se procede a visualizar el tipo de datos de cada columna:

df_Steam_Games.info()

Se aprecia que los tipos de datos no son acordes a la información registrada. Esto será insumo para el ETL correspondiente de este archivo (cambiar tipo de dato correspondiente por columna). A continación visualizamos los datos:

df_Steam_Games.head(5)

Se verifica si 'id' es un identificador único:

# Verificar duplicados en la columna 'id'
duplicados = df_Steam_Games['id'].duplicated()

# Contar la cantidad de duplicados en la columna 'id'
cantidad_duplicados = df_Steam_Games['id'].duplicated().sum()
print(f"Cantidad de duplicados en 'id': {cantidad_duplicados}")

# Eliminar duplicados basados en la columna 'id' en el mismo DataFrame
df_Steam_Games.drop_duplicates(subset=['id'], inplace=True)

# Mostrar el DataFrame resultante sin duplicados
df_Steam_Games.info()


df_Steam_Games.head(3)

### Observaciones:
- Las columnas app_name y title parecen contener la misma información, se revisará en el ETL la posibilidad de imputar valores faltantes entre ellas y eliminar una columna.
- La columna tags contiene información de genre y Price (cuando es Free to Play). Esto se usará para imputar valores en ambas columnas y eliminar tags.
- Url y review url no paracen contener información relevante ni para el modelo de recomendación, ni para los endpoints de la API. Se eliminarán.
- No se eliminarán filas ni se imputarpan datos luego de realizado lo anterior, dado que para cada caso específico se manejarán filtros (consultas) de las columnas relevantes para cada enpoint o el modelo.

A continuación se exporta el datafame a formato CSV, para ser trabajado posteriormente en los ETL con las observaciones anteriores. El archivo se guardará en formato CSV en la carpeta Revised_Data:

# Exportar el DataFrame a CSV
df_Steam_Games.to_csv('Revised_Data/Steam_Games.csv', index=False)

print(f"DataFrame exportado a {'Revised_Data/Steam_Games.csv'}")

### 4. Análisis Preliminar del archivo "user_reviews.json.gz":

La función para abrir este archivo no funcionó bien porque los datos del archivo ".json" no continen el formato adecuado, según se aprecia en el error los datos (Clave - Valor) no están encerrados en doble comillas (") por lo que fué necesario buscar otro código que lea esta estructura. Se encuentra que "ast.literal" nos ayuda a leer cada línea como una expresión literal de python y lo agrega a una lista. Antes de usar, es necesario conocer el "encoding" del archivo para evitar errores de lectura.

# Función para encontrar encoding:

def detect_encoding(file_path):
        
    with open(file_path, 'rb') as f:
        result = chardet.detect(f.read())
        return result['encoding']
    

Se ingresa la dirección del archivo ".json" descomprimido para buscar el encoding del mismo:

# Ejecución de la función para detectar encoding en el archivo:

file_path = 'Original_Data/australian_user_reviews.json'
encoding_UR = detect_encoding(file_path)

print(f"La codificación del archivo es: {encoding_UR}")  

Con el encoding conocido se procede a abrir el archivo en modo lectura para agragar cada fila en "data_list" para finalmente convertirlo en un dataframe:

# Leyendo los datos de "australian_user_reviews.json"
data_list = []

with open ('Original_Data/australian_user_reviews.json', encoding=encoding_UR) as f:
    for line in f.readlines():
        data_list.append(ast.literal_eval(line))

df_User_Review = pd.DataFrame(data_list)
df_User_Review.head()

Se verifica si hay datos nulos en el archivo, encontrando que no hya registros vacíos:

# Muestra la cuenta de NaN por columnas:
print(df_User_Review.isna().sum())
print('----------------------------------------------|')
print(df_User_Review.info())

Se observa que la columna "reviews" contiene datos anidados:

df_User_Review['reviews'].iloc[0]

### Observaciones:
- La columna de url es innecesaria para la realización de los diferentes enpoints y el modelo de recomendación.
- La columna "reviews" debe expandirse conservando el "user_id".

A continuación se exporta el datafame a formato CSV, para ser trabajado posteriormente en los ETL con las observaciones anteriores. El archivo se guardará en formato CSV en la carpeta Revised_Data:

# Exportar el DataFrame a CSV
df_User_Review.to_csv('Revised_Data/User_Review.csv', index=False)

print(f"DataFrame exportado a {'Revised_Data/User_Review.csv'}")

### 4. Análisis Preliminar del archivo "users_items.json.gz":

Del mismo modo que el archivo anterior, los datos separados por comillas sencillas complican que el archivo se abra fácilmente, así que se hace uso de "ast.literal":

# Ejecución de la función para detectar encoding en el archivo:

#file_path = 'Original_Data/australian_users_items.json'
#encoding_UI = detect_encoding(file_path)

#print(f"La codificación del archivo es: {encoding_UI}")  

# Leyendo los datos de "australian_user_reviews.json"
data_list_UI = []

with open ('Original_Data/australian_users_items.json', encoding='MacRoman') as f:
    for line in f.readlines():
        data_list_UI.append(ast.literal_eval(line))

df_User_Items = pd.DataFrame(data_list_UI)
df_User_Items.head()

Se verifica si hay datos nulos en el archivo, encontrando que no hay registros vacíos:

# Muestra la cuenta de NaN por columnas:
print(df_User_Items.isna().sum())
print('----------------------------------------------|')
print(df_User_Items.info())

Se observa que la columna "items" contiene datos anidados:

df_User_Items['items'].iloc[0]

### Observaciones:
- La columna de url es innecesaria para la realización de los diferentes enpoints y el modelo de recomendación.
- La columna "steam_id" parece representar el id del video juego, pero debe revisarse si tiene relación con id en el dataset de Steam_Games porque al parecer en dicho archivo "id" solo representa una parte del "steam_id".
- La columna "items_Count" podría eliminarse ya que de necesitar dicha consulta debería ser reproducible de contar los ítems anidados en la columna "items".
- Se debe extender la columna "items" porque se encuentra anidada.

A continuación se exporta el datafame a formato CSV, para ser trabajado posteriormente en los ETL con las observaciones anteriores. El archivo se guardará en formato CSV en la carpeta Revised_Data:

# Exportar el DataFrame a CSV
df_User_Items.to_csv('Revised_Data/User_Items.csv', index=False)

print(f"DataFrame exportado a {'Revised_Data/User_Items.csv'}")

No se realiza eliminación de registros (filas) más allá de las filas comppletamente vacías, dado que para cada endpoint y el modelo podría tenerse una importancia muy selectiva de las columnas, lo que significa que estas elimnaciones pueden manejarse a traves de filtros o cosnultas específicas para cada función.
