==> 2. ETL.ipynb <==
# Extracción, Transformación y Limpieza de los datos (ETL):

En la carpeta "Revised_Data" se guardan en formato CSV los dataset que deben ser limpiados y transformados para poder crear las fórmulas para los endpoints de la API y el modelo de recomendación.

### 1. Importación de librerías necesarias para realizar los ETL:

# Para abrir los archivos y realizar transformaciones:

import pandas as pd                                         # Para abrir y transformar los datos
import ast                                                  # Para abrir los archivos con datos anidados
from pandas import json_normalize                           # Para expandir los datos anidados

#----------------------------------------------------------------------------------------------------------------------------------------

# Para el modelo de análisis de sentimiento en "df_User_Review":

import re                                                   # Para eliminar caracteres especiales
from nltk.corpus import stopwords                           # para eliminar palabras comunes del ingles
from nltk.tokenize import word_tokenize                     # Para tokenizar
from nltk.stem import WordNetLemmatizer                     # Para lematizar
from nltk import download                                   # Para usar los recursos de nlkt que no están incluídos en la biblioteca
from nltk.sentiment import SentimentIntensityAnalyzer       # Para análisis de sentimiento de "Reviews"
from datetime import datetime                               # Para arreglar el formato de fecha

#----------------------------------------------------------------------------------------------------------------------------------------

# Para ignorar advertencias:

import warnings                    
warnings.filterwarnings("ignore") 



### 2. ETL Steam_Games:

Visualizamos los datos del dataframe "Steam_Games":

df_Steam_Games = pd.read_csv('Revised_Data\Steam_Games.csv')
df_Steam_Games.head()

### 2.1. Se Imputan los valores nulos de "app_name" con "tile" y se elimina la columna "title":

# Llenar los valores vacíos en 'app_name' con los valores de 'title'
df_Steam_Games['app_name'].fillna(df_Steam_Games['title'], inplace=True)

# Eliminar la columna 'title'
df_Steam_Games.drop(columns=['title'], inplace=True)

# Mostrar el DataFrame resultante
df_Steam_Games.head(3)


### 2.2. Se eliminan las columnas "url" y "reviews_url":

# Eliminar las columnas especificadas
columnas_a_eliminar = ['url', 'reviews_url']
df_Steam_Games.drop(columns=columnas_a_eliminar, inplace=True)

# Mostrar el DataFrame resultante
df_Steam_Games.head(3)

### 2.3. Se imputan valores nulos en "genres" según el contenido en "tags"

# Obtener valores únicos de la columna 'tags' para realizar imputaciones de "genres" y "price" cuando es "Free to Play"
valores_unicos_tags = df_Steam_Games['tags'].unique()

# Mostrar los valores únicos
print("Valores únicos en la columna 'tags':")
print(valores_unicos_tags)


# Imputar en 'genres' para valores nulos el contenido en 'tags'
df_Steam_Games['genres'] = df_Steam_Games.apply(lambda row: row['tags'] if pd.isnull(row['genres']) else row['genres'], axis=1)

# Mostrar el DataFrame resultante
df_Steam_Games.head(3)


### 2.4. Convertir datos en el formato correspondiente los datos de cada columna:

# Convertir a tipo de dato string
columnas_a_string = ['publisher', 'genres', 'app_name', 'tags', 'specs', 'developer']
df_Steam_Games[columnas_a_string] = df_Steam_Games[columnas_a_string].astype(str)

# Reemplazar "Free to Play" por 0 en 'price'
df_Steam_Games['price'] = df_Steam_Games['price'].replace('Free to Play', 0)

# Convertir 'release_date' en fecha:
df_Steam_Games['release_date'] = pd.to_datetime(df_Steam_Games['release_date'], errors='coerce')

# Crea la columna 'Year' extrayendo el año y colocando ceros para los valores nulos
df_Steam_Games['release_year'] = df_Steam_Games['release_date'].dt.year.fillna(0).astype(int)

# Convertir 'price' a tipo de dato float
df_Steam_Games['price'] = pd.to_numeric(df_Steam_Games['price'], errors='coerce')

# Reemplazar representaciones específicas de NaN con el valor real de NaN
df_Steam_Games = df_Steam_Games.replace('nan', pd.NA)

# Mostrar las primeras filas del DataFrame
df_Steam_Games.head(3)



### 2.5. Imputar nulos de "price" si en "tags" se tiene "Free to Play" con 0 y eliminar la columna "tags":

# Imputar valores nulos en 'price' si 'tags' contiene 'Free to Play'
df_Steam_Games.loc[df_Steam_Games['price'].isnull() & df_Steam_Games['tags'].str.contains('Free to Play'), 'price'] = 0

# Mostrar las primeras filas del DataFrame
df_Steam_Games[['app_name', 'price']].head()


# Eliminar la columna "tags":
df_Steam_Games.drop(columns='tags', inplace=True)

# Mostrar el DataFrame resultante
df_Steam_Games.head(3)

### 2.5. Exportar el archivo limpio a la carpeta Clean_Data para realizar un EDA en función del modelo y los endpoints:

# Exportar el DataFrame a CSV
df_Steam_Games.to_csv('Clean_Data/SGames_CD.csv', index=False)

print(f"DataFrame exportado a {'Clean_Data/SGames_CD.csv'}")

Observación: Desde el principio se sabe que las columnas "genres" y "specs" tienen datos en listas que deberían extenderse; pero no se realiza en este ETL porque se quiere revisar que para el modelo de recomendación puede ser necesario expandir la lista y aplicar la codificación "One Hot Encoding" para estas variables. Para los enpoints, tal vez sea necesario normalizar las tablas para evitar multiplicar los registros (crear tablas auxiliares "genres" y "specs" que se relaconen con el id del juego).

### 3. ETL Users_Items:

### 3.1. Abrir el archivo usando ast.literal_eval debido a los datos anidados:

# Leer el archivo CSV
df_User_Items = pd.read_csv('Revised_Data\\User_Items.csv')

# Aplicar ast.literal_eval a la columna con diccionarios en listas
df_User_Items['items'] = df_User_Items['items'].apply(ast.literal_eval)

# Mostrar las primeras filas del DataFrame
df_User_Items.head()


### 3.2. Eliminar las columnas 'steam_id','items_count','user_url':

# Eliminar la columnas innecesarias del df:
columnas= ['steam_id','items_count','user_url']
df_User_Items.drop(columns=columnas, inplace=True)

# Hacemos una copia del df para trabajar sobre esta:
df_User_Items_Copy = df_User_Items.copy()

# Mostrar el DataFrame resultante
df_User_Items_Copy.head(3)

### 3.3. Expandir la columna "items" (desanidar):

# Aplicar json_normalize directamente en la columna 'items' y conservar 'user_id':
df_expanded = pd.json_normalize(df_User_Items_Copy.to_dict('records'), 'items', ['user_id'])

# Reorganizar las columnas:
df_User_Items_Copy2 = df_expanded[['user_id', 'item_id', 'item_name', 'playtime_forever', 'playtime_2weeks']]

# Mostrar el nuevo DataFrame:
df_User_Items_Copy2


Verificar que no hayan datos nulos, ya que de haberlos significaría un error en la exploción de "items":

# Muestra la cuenta de NaN por columnas:
print(df_User_Items_Copy2.isna().sum())
print('----------------------------------------------|')
print(df_User_Items_Copy2.info())

Eliminar item_name para optimizar las tablas:

df_User_Items_Copy2.drop(columns=['item_name'], inplace=True)
df_User_Items_Copy2

df_User_Items_Copy2.iloc[88297]

### 3.4. Exportar el archivo limpio a la carpeta Clean_Data para realizar un EDA en función del modelo y los endpoints:

# Exportar el DataFrame a CSV
df_User_Items_Copy2.to_csv('Clean_Data/UItems_CD.csv', index=False)

print(f"DataFrame exportado a {'Clean_Data/UItems_CD.csv'}")

### 4. ETL Users_Review:

### 4.1. Abrir el archivo usando ast.literal_eval debido a los datos anidados:

# Leer el archivo CSV
df_User_Review = pd.read_csv('Revised_Data\\User_Review.csv')

# Aplicar ast.literal_eval a la columna con diccionarios en listas
df_User_Review['reviews'] = df_User_Review['reviews'].apply(ast.literal_eval)

# Mostrar las primeras filas del DataFrame
df_User_Review.head()

### 4.2. Eliminar la columna 'user_url':

# Eliminar la columna "User_url":
df_User_Review.drop(columns=['user_url'], inplace=True)

# Hacemos una copia del df para trabajar sobre esta:
df_User_Review_Copy = df_User_Review.copy()

# Mostrar el DataFrame resultante
df_User_Review_Copy.head(3)

### 4.3. Expandir la columna "reviews" (desanidar):

Buscar las claves de los datos que servirán como columnas:

# Imprimir el contenido de la columna 'reviews' (solo la primera fila) para obtener claves:
ejemplo_dato=df_User_Review_Copy['reviews'].iloc[:1]
ejemplo_dato=dict(ejemplo_dato)

# Obtener las claves de los diccionarios en la lista
claves = [list(diccionario.keys()) for diccionario in ejemplo_dato[0]]

# Imprimir las claves
print(claves[0])


Expandir datos:

# Aplicar json_normalize directamente en la columna 'reviews' y conservar 'user_id':
df_expanded2 = pd.json_normalize(df_User_Review_Copy.to_dict('records'), 'reviews', ['user_id'])

# Reorganizar las columnas:
df_User_Review_Copy2 = df_expanded2[['user_id','funny', 'posted', 'last_edited', 'item_id', 'helpful', 'recommend', 'review']]

# Mostrar el nuevo DataFrame:
df_User_Review_Copy2

Verificar que no hayan datos nulos, ya que de haberlos significaría un error en la exploción de "reviews":

# Muestra la cuenta de NaN por columnas:
print(df_User_Review_Copy2.isna().sum())
print('----------------------------------------------|')
print(df_User_Review_Copy2.info())

### 4.4. Limpieza y normalización de "reviews" para procesarlos antes de realizar análisis de sentimiento:

# Descargar los recursos necesarios (puedes ejecutarlo una vez)
download('punkt')
download('stopwords')
download('wordnet')

# Función para limpiar, tokenizar, normalizar y lematizar el texto
def preprocess_text(text):
    # Eliminar caracteres especiales y convertir a minúsculas
    text = re.sub(r'[^a-zA-Z\s]', '', str(text).lower())
    
    # Tokenizar
    words = word_tokenize(text)
    
    # Eliminar stopwords
    stop_words = set(stopwords.words('english'))
    words = [word for word in words if word not in stop_words]
    
    # Lematizar
    lemmatizer = WordNetLemmatizer()
    words = [lemmatizer.lemmatize(word) for word in words]
    
    # Unir las palabras procesadas en un string
    processed_text = ' '.join(words)
    
    return processed_text

# Aplicar la función de preprocesamiento a la columna 'review'
df_User_Review_Copy2['processed_review'] = df_User_Review_Copy2['review'].apply(preprocess_text)

# Mostrar el DataFrame con la nueva columna procesada
print(df_User_Review_Copy2[['review', 'processed_review']])


### 4.5. Aplicar análisis de sentimiento a la nueva columna procesada review "processed_reviews":

import nltk
from nltk.sentiment import SentimentIntensityAnalyzer

# Descargar el recurso necesario (puedes ejecutarlo una vez)
nltk.download('vader_lexicon')

# Crear una instancia de SentimentIntensityAnalyzer
sid = SentimentIntensityAnalyzer()

# Función para asignar el sentimiento basado en compound score
def analyze_sentiment(text):
    sentiment_score = sid.polarity_scores(str(text))['compound']
    
    if sentiment_score >= 0.05:
        return 2  # Positivo
    elif sentiment_score <= -0.05:
        return 0  # Negativo
    else:
        return 1  # Neutral

# Aplicar la función a la columna 'review' y crear la nueva columna 'sentiment_analysis'
df_User_Review_Copy2['sentiment_analysis'] = df_User_Review_Copy2['processed_review'].apply(analyze_sentiment)

# Mostrar el DataFrame resultante
df_User_Review_Copy2


### 4.6. Eliminar "review" y renombrar "processed_review" para dejar solo una columan definitiva:

# Eliminar la columna 'review'
df_User_Review_Copy2.drop('review', axis=1, inplace=True)

# Renombrar 'processed_review' como 'review'
df_User_Review_Copy2.rename(columns={'processed_review': 'review'}, inplace=True)

# Mostrar el DataFrame resultante
df_User_Review_Copy2.head(3)


### 4.7. Crear una columna año que servirá para las consultas para los enpoints:

# Crear una columna auxiliar 'Año' que extrae el año después de la coma en 'posted'
df_User_Review_Copy2['posted_year'] = df_User_Review_Copy2['posted'].str.extract(r', (\d{4})')

# Elimniar datos nulos y la columna posted:
df_User_Review_Copy2.dropna(subset=['posted'], inplace=True)
df_User_Review_Copy2.dropna(subset=['posted_year'], inplace=True)
df_User_Review_Copy2.drop('posted', axis=1, inplace=True)

# Mostrar el DataFrame resultante
print(df_User_Review_Copy2[['posted_year']].head(3))


# Cambiar formato de Año a número:
df_User_Review_Copy2['posted_year'] = df_User_Review_Copy2['posted_year'].astype('Int16')

# Verificar formato de Año:
print(df_User_Review_Copy2.info()) 

# Imprimir el DataFrame actualizado
df_User_Review_Copy2.head(3)

### 4.8. Eliminar "last_edited":

df_User_Review_Copy2.drop(columns=['last_edited'], inplace=True)

### 4.9. Eliminar "Review" porque ya se ha capturado su importancia en el análisis de sentimiento y helpful ya que no haremos uso de esta columna:

df_User_Review_Copy2.drop(columns=['review'], inplace=True)
df_User_Review_Copy2.drop(columns=['helpful'], inplace=True)
df_User_Review_Copy2.drop(columns=['funny'], inplace=True)

### 4.10. Exportar el archivo limpio a la carpeta Clean_Data para realizar un EDA en función del modelo y los endpoints:

# Exportar el DataFrame a CSV
df_User_Review_Copy2.to_csv('Clean_Data/UReviews_CD.csv', index=False)

print(f"DataFrame exportado a {'Clean_Data/UReviews_CD.csv'}")
