==> 4. Funciones_Endpoints.ipynb <==
# Creación de Consultas para los Endpoints de la API:

En este notebook se busca crear las consultas (realizar la agregación de datos) necesarios para que las funciones se ejecuten en la aplicación. También se crean y se prueban las diferentes fucniones:

#### 1. Importar librerías y archivos necesarios:

import pandas as pd
import numpy as np

import warnings                    
warnings.filterwarnings("ignore") 

# Steam Games:
SGames = pd.read_csv('Clean_Data\\SGames_CD.csv', quotechar='"')

# Users Items:
UItems = pd.read_csv('Clean_Data\\UItems_CD.csv', quotechar='"')

# Users Reviews:
UReviews = pd.read_csv('Clean_Data\\UReviews_CD.csv', quotechar='"')

# Users Reviews:
Genres_Items = pd.read_csv('Clean_Data\Genres_Items.csv', quotechar='"')

### 2. Crear la formula registros por desarrollador:

def developer( desarrollador : str ): Cantidad de items y porcentaje de contenido Free por año según empresa desarrolladora. Ejemplo de retorno:

| Año  | Cantidad de artículos | Contenido Gratis |
|------|------------------------|-------------------|
| 2023 | 50                     | 27%               |
| 2022 | 45                     | 25%               |
| xxxx | xx                     | x%                |


2.1. Crear datos de entrada:

# Seleccionar las columnas deseadas y eliminar filas con datos nulos
End1_Developer = SGames[['id', 'developer', 'release_year', 'price']].dropna()
End1_Developer = End1_Developer.drop(End1_Developer[End1_Developer['release_year'] == 0].index) # así si
#End1_Developer = SGames[['id', 'developer', 'release_year', 'price']].replace(0, np.nan).dropna(subset=['release_year']) #ojo así no!



# Mostrar el DataFrame resultante
print("\nDataFrame final:")
End1_Developer


2.2. Obtener top-five de desarrolladores con más items para probar la fórmula:

# Calcular el top 5 de desarrolladores que más aparecen
top_developers = End1_Developer['developer'].value_counts().head(-5)

# Mostrar el top 5
print("Top 5 de desarrolladores que más aparecen:")
print(top_developers)

2.3. Crear fórmula "get_developer_stats":

def get_developer_stats(developer):
    
    # Filtrar el DataFrame por desarrollador
    developer_df = End1_Developer[End1_Developer['developer'] == developer]

    # Calcular la cantidad de items y el porcentaje de contenido gratuito por año
    stats_by_year = developer_df.groupby('release_year').agg({'id': 'count', 'price': lambda x: (x == 0).sum() / len(x) * 100}).reset_index()
   
    stats_by_year.columns = ['release_year', 'item_count', 'free_content_percentage']

    # Convertir los resultados a una lista de diccionarios
    result_list = stats_by_year.to_dict(orient='records')

    return result_list



2.4. Probar la fórmula:

# Ejemplo de uso de la función
print(get_developer_stats('Secret Level SRL'))

print(get_developer_stats('KOEI TECMO GAMES CO., LTD.'))

2.5. Guardar consulta (insumo para la API) en la carperta Data_Queries:

# Exportar el DataFrame a CSV
End1_Developer.to_csv('Data_Queries/End1_Developer.csv', index=False)

print(f"DataFrame exportado a {'Data_Queries/End1_Developer.csv'}")

### 3. Crear función registros de usuario:

def userdata( User_id : str ): Debe devolver cantidad de dinero gastado por el usuario, el porcentaje de recomendación en base a reviews.recommend y cantidad de items.


Ejemplo de retorno: {"Usuario X" : us213ndjss09sdf, "Dinero gastado": 200 USD, "% de recomendación": 20%, "cantidad de items": 5}

3.1. Crear datos de entrada:


# Crear el DataFrame End2_User
UItems_F= UItems[['user_id', 'item_id']]
End1_Developer_F=End1_Developer[['id','price']]
End2_User = pd.merge(UItems_F, End1_Developer, left_on='item_id', right_on='id')
End2_User = End2_User[['id', 'user_id', 'price']]
UReviews_F = UReviews[['item_id', 'user_id','recommend']]

# Agrupar End2_User en total de dienero gastado por usuario:
End2_User = End2_User.groupby('user_id')['price'].sum().reset_index()

# Mostrar el DataFrame resultante
print("\nDataFrame End2_User:")
print(End2_User.head(-5))

print("\nDataFrame UReviews_F:")
print(UReviews_F.head(-5))

3.2. Crear fórmula "userdata":

def userdata(User_id):
    # Filtrar el DataFrame End2_User por user_id
    user_df = End2_User[End2_User['user_id'] == User_id]

    # Calcular la cantidad de dinero gastado por el usuario
    total_money_spent = user_df['price'].sum()

    # Calcular la cantidad de items
    total_items = len(user_df)

    # Filtrar el DataFrame UReviews_F por item_ids relacionados con el usuario
    user_reviews_df = UReviews_F[UReviews_F['user_id'] == User_id]

    # Calcular el porcentaje de recomendación basado en elementos únicos
    unique_items_count = user_reviews_df['item_id'].nunique()
    unique_items_with_recommendation = user_reviews_df[user_reviews_df['recommend'] == True]['item_id'].nunique()
    recommendation_percentage = (unique_items_with_recommendation / unique_items_count) * 100 if unique_items_count > 0 else 0

    # Crear el diccionario de resultados
    result_dict = {
        "Usuario": User_id,
        "Dinero gastado": f"{total_money_spent} USD",
        "% de recomendación": f"{recommendation_percentage:.2f}%",
        "Cantidad de items": total_items
    }

    return result_dict


3.3. Probar fórmula:

# Ejemplo de uso de la función
user_stats_result = userdata('76561197970982479')
print(user_stats_result)

3.4. Guardar consulta (insumo para la API) en la carperta Data_Queries:

# Exportar el DataFrame a CSV:

End2_User.to_parquet('Data_Queries/End2_User.parquet', index=False)
print(f"DataFrame exportado a {'Data_Queries/End2_User.parquet'}")

UReviews_F.to_csv('Data_Queries/UReviews_F.csv', index=False)
print(f"DataFrame exportado a {'Data_Queries/UReviews_F.csv'}")

### 4. Función mejor Usuario por Género de video juego:

def UserForGenre( genero: str ) : Debe devolver el usuario que acumula más horas jugadas para el género dado y una lista de la acumulación de horas jugadas por año de lanzamiento.

Ejemplo de retorno: {"Usuario con más horas jugadas para Género X" : us213ndjss09sdf, "Horas jugadas":[{Año: 2013, Horas: 203}, {Año: 2012, Horas: 100}, {Año: 2011, Horas : 23}]}

4.1. Crear datos de Entrada:

# crear subset de SGames con las columnas necesarias:
subset_SGames = SGames[['id', 'release_year']]

# Unir subset_SGames y Genres_Items por la columna 'id':
merged_data = pd.merge(subset_SGames, Genres_Items, on='id')
merged_data

# Seleccionar las columnas 'user_id', 'item_id' y 'playtime_forever' de UItems_CD:
UItems_subset = UItems[['user_id', 'item_id', 'playtime_forever']]

# Unir UItems_CD_subset con merged_data por las columnas 'item_id' e 'id':
End3_Genre_User = pd.merge(UItems_subset, merged_data, left_on='item_id', right_on='id')

# Eliminar la columna duplicada 'id':
End3_Genre_User = End3_Genre_User.drop(columns=['id'])

# Eliminar filas con Año==0 en UserForGenre:
End3_Genre_User = End3_Genre_User[End3_Genre_User['release_year'] != 0]

# Mostrar el DataFrame resultante:
End3_Genre_User.head()


Se encuentra que los datos son demasiado grandes para guardarlos así que se procede a crear una dataframe agregado:

# Agrupa por género y usuario, y calcula la suma de playtime_forever
grouped_data = End3_Genre_User.groupby(['genres', 'user_id']).agg({'playtime_forever': 'sum'}).reset_index()

# Encuentra el usuario con la máxima suma de playtime_forever en cada género
max_playtime_users = grouped_data.groupby('genres')['playtime_forever'].idxmax()
max_playtime_data = grouped_data.loc[max_playtime_users]

# Ahora, agrupa por usuario y año y calcula la suma de playtime_forever
user_year_data = End3_Genre_User.groupby(['user_id', 'release_year']).agg({'playtime_forever': 'sum'}).reset_index()

# Fusiona los datos de max_playtime_data con user_year_data
End3_Genre_User = pd.merge(max_playtime_data, user_year_data, on='user_id', how='left')

# Muestra el DataFrame resultante
End3_Genre_User

4.2. Crear función UserForGenre:

def UserForGenre(genero: str):
    # Filtrar UserForGenre por el género dado
    genre_data = End3_Genre_User[End3_Genre_User['genres'] == genero]

    # Encontrar al usuario con más horas jugadas para el género dado
    max_hours_user = genre_data.loc[genre_data['playtime_forever_x'].idxmax()]['user_id']

    # Crear una lista de la acumulación de horas jugadas por año de lanzamiento
    hours_by_year = genre_data.groupby('release_year')['playtime_forever_y'].sum().reset_index()
    hours_by_year_list = [{"Año": int(year), "Horas": int(hours)} for year, hours in hours_by_year.values]

    # Crear el diccionario de resultados
    result_dict = {
        f"Usuario con más horas jugadas para el género {genero}": max_hours_user,
        "Horas jugadas": hours_by_year_list
    }

    return result_dict


4.3. Probar función:

# Ejemplo de uso
genero_ejemplo = "Action"  # Reemplazar con el género deseado
resultado = UserForGenre(genero_ejemplo)
print(resultado)

4.5. Guardar consulta (insumo para la API) en la carperta Data_Queries:

# Exportar el DataFrame a CSV:

End3_Genre_User.to_parquet('Data_Queries/End3_Genre_User.parquet', index=False)
print(f"DataFrame exportado a {'Data_Queries/End3_Genre_User.parquet'}")


# Agrupa por género y usuario, y calcula la suma de playtime_forever
grouped_data = End3_Genre_User.groupby(['genres', 'user_id']).agg({'playtime_forever': 'sum'}).reset_index()

# Encuentra el usuario con la máxima suma de playtime_forever en cada género
max_playtime_users = grouped_data.groupby('genres')['playtime_forever'].idxmax()
max_playtime_data = grouped_data.loc[max_playtime_users]

# Ahora, agrupa por usuario y año y calcula la suma de playtime_forever
user_year_data = End3_Genre_User.groupby(['user_id', 'release_year']).agg({'playtime_forever': 'sum'}).reset_index()

# Fusiona los datos de max_playtime_data con user_year_data
End3_Genre_User = pd.merge(max_playtime_data, user_year_data, on='user_id', how='left')

# Muestra el DataFrame resultante
End3_Genre_User

### 5. Crear función de mejor desarrollador por año:

def best_developer_year( año: int ) : Devuelve el top 3 de desarrolladores con juegos MÁS recomendados por usuarios para el año dado. (reviews.recommend = Verdadero y comentarios positivos)

Ejemplo de retorno: [{"Puesto 1" : X}, {"Puesto 2" : Y},{"Puesto 3" : Z}]


5.1. Crear los datos de entrada:

# Filtrar por recommend == True y sentiment_analysis == 2
filtered_reviews = UReviews[(UReviews['recommend'] == True) & (UReviews['sentiment_analysis'] == 2)]

# Resumir por el conteo de "item_id" por cada año "posted_year"
UReviews_gb = filtered_reviews.groupby(['posted_year', 'item_id']).size().reset_index(name='count')

# Mostrar el DataFrame resultante
UReviews_gb


Solo se requiere saber el desarrollador y para ello se debe filtrar el dataframe de "SGames" y unir con la tabala anterior:

# Crear subconjunto de "SGames":
Games_subset= SGames[['developer','id']]

#Unir con la tabla "UReviews_gb"
End4_BestDeveloperYear= pd.merge(Games_subset, UReviews_gb, left_on='id', right_on='item_id')

# Seleccionar las columnas necesarias:
End4_BestDeveloperYear=End4_BestDeveloperYear[['developer', 'item_id','posted_year','count']]
End4_BestDeveloperYear

5.2. Crear función "best_developer_year":

def best_developer_year(año: int):
    # Filtrar End4_BestDeveloperYear por el año dado
    filtered_data = End4_BestDeveloperYear[End4_BestDeveloperYear['posted_year'] == año]

    # Filtrar por juegos más recomendados (recommend = Verdadero y comentarios positivos)
    top_developers = filtered_data[filtered_data['count'] > 0]

    # Sumar los conteos por desarrollador y obtener el ranking de los 3 mejores desarrolladores
    top3_developers = top_developers.groupby('developer')['count'].sum().nlargest(3).reset_index()

    # Crear la lista de resultados en el formato deseado
    result_list = [{"Puesto {}".format(i + 1): developer} for i, developer in enumerate(top3_developers['developer'])]

    return result_list

5.3. Probar función:

# Ejemplo de uso
año_ejemplo = 2015  # Reemplazar con el año deseado
resultado = best_developer_year(año_ejemplo)
print(resultado)

End4_BestDeveloperYear.info()

End4_BestDeveloperYear = End4_BestDeveloperYear.dropna(subset=['developer']) # Eliminar nulos por developer no tiene sentido dejarlos

5.4. Guardar consulta (insumo para la API) en la carperta Data_Queries:

# Exportar el DataFrame a CSV:

End4_BestDeveloperYear.to_csv('Data_Queries/End4_BestDeveloperYear.csv', index=False)
print(f"DataFrame exportado a {'Data_Queries/End4_BestDeveloperYear.csv'}")

### 6. Crear función que mida la percepción por desarrollador:

def desarrollador_reviews_analysis( desarrolladora: str ) : Según el desarrollador, se devuelve un diccionario con el nombre del desarrollador como llave y una lista con la cantidad total de registros de reseñas de usuarios que se encuentran categorizados con un análisis de sentimiento como valor positivo o negativo.

Ejemplo de retorno: {'Valve' : [Negative = 182, Positive = 278]}

6.1. Crear los datos de entrada:

# crear subconjunto de UReviews con el análisis de sentimiento:
sentimiento=UReviews[['item_id','sentiment_analysis']]

# Unimos data set creutilizando el subconjunto creado de Games_subset:
End5_Sentimiento=pd.merge(Games_subset, sentimiento, left_on='id', right_on='item_id')

# Seleccionamos columnas necesarias:
End5_Sentimiento=End5_Sentimiento[['developer','item_id','sentiment_analysis']]

# Revisamos dataframe:
End5_Sentimiento


# Eliminar filas con valores nulos en cualquier columna
End5_Sentimiento = End5_Sentimiento.dropna()

# Reconstruir tabla para mejorar la lectura en los endpoints:

# Filtrar solo las reseñas positivas y negativas
positivas = End5_Sentimiento[End5_Sentimiento['sentiment_analysis'] == 2]
negativas = End5_Sentimiento[End5_Sentimiento['sentiment_analysis'] == 0]

# Contar el número de reseñas positivas y negativas por desarrollador
positivas_counts = positivas.groupby('developer').size().reset_index(name='Positive')
negativas_counts = negativas.groupby('developer').size().reset_index(name='Negative')

# Combinar los resultados en un nuevo DataFrame
resultado_final = pd.merge(positivas_counts, negativas_counts, on='developer', how='outer').fillna(0)

End5_Sentimiento = resultado_final

End5_Sentimiento

6.2. Crear función "developer_reviews_analysis":

def desarrollador_reviews_analysis(ReviewsDeveloper: str):
   
    # Filtrar End5_Sentimiento por el desarrollador dado
    desarrollador_data = End5_Sentimiento[End5_Sentimiento['developer'] == ReviewsDeveloper]

    # Verificar si el desarrollador existe en el DataFrame
    if not desarrollador_data.empty:
        # Obtener la cantidad de reseñas positivas y negativas
        negative_count = int(desarrollador_data['Negative'])
        positive_count = int(desarrollador_data['Positive'])

        # Crear el diccionario de resultados
        result_dict = {ReviewsDeveloper: {'Negative': negative_count, 'Positive': positive_count}}

        return result_dict
    else:
        return {ReviewsDeveloper: {'Negative': 0, 'Positive': 0}}


6.3. Probar función:

# Ejemplo de uso
desarrolladora_ejemplo = 'Valve'  # Reemplazar con el nombre del desarrollador deseado
resultado = desarrollador_reviews_analysis(desarrolladora_ejemplo)
print(resultado)

6.4. Guardar consulta (insumo para la API) en la carperta Data_Queries:

# Exportar el DataFrame a CSV:

End5_Sentimiento.to_csv('Data_Queries/End5_Sentimiento.csv', index=False)
print(f"DataFrame exportad  o a {'Data_Queries/End5_Sentimiento.csv'}")
