==> 5. Modelo_Recomendación.ipynb <==
# Modelo de Recomendación Item-Item:


El modelo debe establecer una relación ítem-ítem, donde la recomendación de elementos similares se basa en la medida de similitud entre un ítem dado y el resto. En este contexto, el modelo toma como entrada un juego específico y genera como salida una lista de juegos recomendados. Para lograr esto, se recomienda aplicar la métrica de similitud del coseno.

### 1. Importar las librerías necesarias para la creación del modelo y descarga de datos:

import pandas as pd                                                 # Para abrir y editar los archivos 
import numpy as np                                                  # Para realizar cálculos

#----------------------------------------------------------------------------------------------------------------------------------

from sklearn.feature_extraction.text import CountVectorizer         # Para convertir las características analizadas en matrices
from sklearn.metrics.pairwise import cosine_similarity              # Para calcular la similitud de coseno de la matriz


### 2. Abrir y arreglar los datos para que no hayan registros nulos:

# Abrir arcchivo "Genres_Items"
df_genres_expanded= pd.read_csv('Clean_Data/Genres_Items.csv')

# Eliminar registros nulos en 'id' o 'genre'
df_genres_expanded = df_genres_expanded.dropna(subset=['id', 'genres'])

df_genres_expanded.info()

### 3. Expandir la columna "genres":

data=df_genres_expanded

# Convertir las listas en cadenas separadas por comas
data['genres'] = df_genres_expanded['genres'].astype(str)

# Crear columnas para cada género con 1 y 0
data = pd.get_dummies(data['genres'], prefix='', prefix_sep='')

# Agregar la columna 'id' al nuevo DataFrame
data['id'] = df_genres_expanded['id'].abs()

# Agrupar por 'id' y tomar el máximo para combinar las filas duplicadas
data = data.groupby('id').max()

# Resetear el índice
data = data.reset_index()

data

4. Abrir el archivo SGames_CD, eliminar las columnas no necesarias y los datos vacíos:

# Abrir arcchivo "Genres_Items"
Model_Data = pd.read_csv('Clean_Data\\SGames_CD.csv', quotechar='"')

#Usar las columnas necesarias
Model_Data=Model_Data[['id','genres','publisher','release_date','specs','price','early_access','developer','release_year']]

# Eliminar registros con valores nulos en 'genres' y valores de 0 en 'release_year'

Model_Data = Model_Data[(Model_Data['release_year'] != 0) & (Model_Data['genres'].notnull())]

# Reemplazar valores nulos en 'genres' con 'desconocido'
Model_Data['genres'] = Model_Data['genres'].fillna('a stranger')

# Reemplazar valores nulos en 'developer' y 'publisher' con 'desconocido':
Model_Data['developer'] = Model_Data['developer'].fillna('a stranger')
Model_Data['publisher'] = Model_Data['publisher'].fillna('a stranger')

# Colocar 0 en los valores nulos de 'price'
Model_Data['price'] = Model_Data['price'].fillna(0)

# Eliminar registros con valores nulos en 'genres'
Model_Data = Model_Data.dropna(subset=['genres'])

# Eliminar la columna 'release_date' y 'genres'
Model_Data = Model_Data.drop(columns=['release_date','genres'])

# Colocar 'not specified' en los valores nulos de la columna 'specs'
Model_Data['specs'] = Model_Data['specs'].fillna('not specified')

# Verificar los cambios
Model_Data.head(-5)


### 5. Unir las tablas para el modelo de recomendación:

df_model_data = pd.DataFrame(Model_Data)

# Merge (join) de los DataFrames por la columna 'id'
result = pd.merge(df_model_data, data, on='id')

# Eliminar la columna 'id' del DataFrame resultante
#result = result.drop('id', axis=1)

# Mostrar el resultado
result

### 6. Convertir "specs" en una cadena sin carácteres, ni separadores para facilitar la lectura del modelo:

Model_Data=result

# Eliminar comillas, corchetes, comas y reemplazar guiones por espacios en 'specs'
Model_Data['specs'] = Model_Data['specs'].astype(str).str.replace('[', '').str.replace(']', '').str.replace("'", "").str.replace(",", "").str.replace("-", " ")

# Verificar los cambios
Model_Data.head()

Model_Data.info()

### 7. Desarrollo del modelo similitud del coseno (item-item):

# Seleccionar solo las columnas numéricas para el cálculo de similitud del coseno
numeric_columns = ['price', 'release_year']

# Normalizar las columnas numéricas
normalized_data = Model_Data[numeric_columns]
normalized_data = (normalized_data - normalized_data.mean()) / (normalized_data.max() - normalized_data.min())

# Concatenar las columnas normalizadas con las columnas categóricas
normalized_model_data = pd.concat([Model_Data[['id', 'publisher', 'specs', 'early_access', 'developer']], normalized_data], axis=1)

# Calcular la similitud del coseno
cosine_sim = cosine_similarity(normalized_data)

### 8. Creación de fórmula para que reciba un "id" de video juego y devuelva los 5 más parecidos a el, para el modelo de recomendación:

# Función para obtener las recomendaciones
def get_recommendations(game_id, n=5):
    # Obtener la fila correspondiente al juego seleccionado
    selected_game_row = normalized_model_data[normalized_model_data['id'] == game_id]

    # Obtener las similitudes del coseno con respecto al juego seleccionado
    similarities = cosine_sim[selected_game_row.index[0]]

    # Obtener los índices de los juegos más similares (excluyendo el juego seleccionado)
    similar_indices = similarities.argsort()[:-1][::-1][:n]

    # Obtener los ID de los juegos más similares
    recommended_ids = normalized_model_data.iloc[similar_indices]['id'].tolist()

    return recommended_ids


# Ejemplo de uso: Obtener las 5 recomendaciones para el juego con ID 1
recommended_games = get_recommendations(772540, n=5)
print(recommended_games)

### 9. Iterar en la base de datos, para obtener para cada id los juesgos recomendados:

dic_similitud = {}

for i in Model_Data['id']:
    try:
        value = get_recommendations(i)
        if value == i:
            continue
        else:
            dic_similitud[i] = value
    except Exception as e:
        print(f"Error para el ID {i}: {str(e)}")

### 10. Crear dataframe del diciconario obtenido:

# Crear listas para almacenar datos
ids = []
recommended_ids = []

# Iterar sobre el diccionario y crear las listas
for key, values in dic_similitud.items():
    ids.extend([key] * len(values))
    recommended_ids.extend(values)

# Crear un DataFrame a partir de las listas
similitud_df = pd.DataFrame({'id': ids, 'recommended_id': recommended_ids})

# Imprimir el DataFrame resultante
similitud_df

# Agrupar por 'id' y concatenar 'recommended_id' en una lista
grouped_df = similitud_df.groupby('id')['recommended_id'].agg(list).reset_index()

grouped_df


### 11. Exportar datos para ser usados por FastApi:

11.1. Crear tabla auxiliar con los nombres de videojuegos:

names = pd.read_csv('Clean_Data\\SGames_CD.csv', quotechar='"')
names=names[['id','app_name']]

# Exportar el DataFrame a CSV:

grouped_df.to_csv('Data_Queries/similitud_df.csv', index=False)
print(f"DataFrame exportado a {'Data_Queries/similitud_df.csv'}")

names.to_csv('Data_Queries/names.csv', index=False)
print(f"DataFrame exportado a {'Data_Queries/names.csv'}")

similitud_df = pd.read_csv('Data_Queries\similitud_df.csv')
#df_genres_expanded= pd.read_csv('Clean_Data/Genres_Items.csv')
  
similitud_df

### 12. Crear Función para FastApi:

def obtener_nombres_recomendados(id_seleccionado):
    try:
        # Obtener el nombre del juego seleccionado
        nombre_seleccionado = names.loc[names['id'] == id_seleccionado, 'app_name'].iloc[0]

        # Obtener los IDs de los juegos recomendados
        juegos_recomendados_ids = similitud_df.loc[similitud_df['id'] == id_seleccionado, 'recommended_id'].iloc[0]

        # Obtener los nombres de los juegos recomendados
        nombres_recomendados = []
        for id_recomendado in juegos_recomendados_ids:
            nombre_recomendado = names.loc[names['id'] == id_recomendado, 'app_name'].iloc[0]
            nombres_recomendados.append(nombre_recomendado)

        # Crear y devolver el diccionario
        diccionario_resultado = {'nombre_seleccionado': nombre_seleccionado, 'nombres_recomendados': nombres_recomendados}
        return diccionario_resultado

    except IndexError:
        print(f"No se encontró nombre para el ID seleccionado {id_seleccionado}")

    except KeyError:
        print(f"No se encontraron recomendaciones para el ID seleccionado {id_seleccionado}")

    except Exception as e:
        print(f"Error inesperado: {str(e)}")

12.1. Probar Función:

# Ejemplo de uso
id_seleccionado = 670290  # Ingresa el ID deseado
resultado = obtener_nombres_recomendados(id_seleccionado)
print(resultado)
